adds context api
