+++ src/components/Cart/Mycart.js
@@ -1,39 +1,16 @@
-import React from 'react';
-import Header from './Header';
-import Main from './Main';
-import Basket from './Basket';
-import data from '../../data';
-import { useState } from 'react';
-import './Kart.css';
+import React, { useContext, useState } from "react";
+import Header from "./Header";
+import Main from "./Main";
+import Basket from "./Basket";
+import data from "../../data";
+import { AppContext } from "../Context/Context";
 
+import "./Kart.css";
 
 function Cart() {
   const { products } = data;
-  const [cartItems, setCartItems] = useState([]);
-  const onAdd = (product) => {
-    const exist = cartItems.find((x) => x.id === product.id);
-    if (exist) {
-      setCartItems(
-        cartItems.map((x) =>
-          x.id === product.id ? { ...exist, qty: exist.qty + 1 } : x
-        )
-      );
-    } else {
-      setCartItems([...cartItems, { ...product, qty: 1 }]);
-    }
-  };
-  const onRemove = (product) => {
-    const exist = cartItems.find((x) => x.id === product.id);
-    if (exist.qty === 1) {
-      setCartItems(cartItems.filter((x) => x.id !== product.id));
-    } else {
-      setCartItems(
-        cartItems.map((x) =>
-          x.id === product.id ? { ...exist, qty: exist.qty - 1 } : x
-        )
-      );
-    }
-  };
+  const [cartItems, onAdd, onRemove] = useContext(AppContext);
+  
   return (
     <div className="App">
       <Header countCartItems={cartItems.length}></Header>
