+++ src/components/Context/Context.js
@@ -1,0 +1,38 @@
+import React, {createContext, useState} from 'react'
+
+export const AppContext = createContext()
+
+export function Context(props) {
+    const [cartItems, setCartItems] = useState([])
+
+    const onAdd = (product) => {
+        const exist = cartItems.find((x) => x.id === product.id);
+        if (exist) {
+          setCartItems(
+            cartItems.map((x) =>
+              x.id === product.id ? { ...exist, qty: exist.qty + 1 } : x
+            )
+          );
+        } else {
+          setCartItems([...cartItems, { ...product, qty: 1 }]);
+        }
+      };
+
+      const onRemove = (product) => {
+        const exist = cartItems.find((x) => x.id === product.id);
+        if (exist.qty === 1) {
+          setCartItems(cartItems.filter((x) => x.id !== product.id));
+        } else {
+          setCartItems(
+            cartItems.map((x) =>
+              x.id === product.id ? { ...exist, qty: exist.qty - 1 } : x
+            )
+          );
+        }
+      };
+    return (
+        <AppContext.Provider value={[cartItems, onAdd, onRemove]}>
+            {props.children}
+        </AppContext.Provider>
+    )
+}
